refactor(mrow): move converter and LinearSystemPattern into mrow folder
