+++ src/data/usecases/mathml-to-latex-convertion/converters/mrow/index.ts
@@ -1,0 +1,1 @@
+export { MRow } from './mrow';
